refactor(protocol-designer): refactor labware and deck components for clarity
* rename PD's LabwareContainer to LabwareOnDeck/index.js (redux)
* rename Component Library's LabwareContainer to LabwareWrapper (non-redux)
* reduce prop-passing-down in LabwareOnDeck via render-prop
* split out DragPreviewLayer from LabwareOnDeck
* Rename ConnectedDeckSetup to DeckSetup/index.js
