+++ protocol-designer/src/components/labware/LabwareOnDeck/constants.js
@@ -1,0 +1,5 @@
+// @flow
+// TODO: BC 2019-01-03 consolidate with DraggableStepItems DND_TYPES
+export const DND_TYPES: { LABWARE: 'LABWARE' } = {
+  LABWARE: 'LABWARE',
+}
