+++ protocol-designer/src/containers/ConnectedMainPanel.js
@@ -3,20 +3,31 @@ import * as React from 'react'
 import { connect } from 'react-redux'
 import { Splash } from '@opentrons/components'
 
-import ConnectedDeckSetup from '../containers/ConnectedDeckSetup'
+import { START_TERMINAL_ITEM_ID, type TerminalItemId } from '../steplist'
+import { Portal as MainPageModalPortal } from '../components/portals/MainPageModalPortal'
+import DeckSetup from '../components/DeckSetup'
 import ConnectedFilePage from '../containers/ConnectedFilePage'
 import SettingsPage from '../components/SettingsPage'
 import LiquidsPage from '../components/LiquidsPage'
+import Hints from '../components/Hints'
+import LiquidPlacementModal from '../components/LiquidPlacementModal.js'
+import LabwareSelectionModal from '../components/LabwareSelectionModal'
+import StepEditForm from '../components/StepEditForm'
+import TimelineAlerts from '../components/alerts/TimelineAlerts'
 
-import type { BaseState } from '../types'
+import { selectors as stepsSelectors } from '../ui/steps'
+import { selectors as labwareIngredSelectors } from '../labware-ingred/selectors'
 import { selectors, type Page } from '../navigation'
+import type { BaseState } from '../types'
 
-type Props = { page: Page }
-
-export default connect<Props, {||}, _, _, _, _>(mapStateToProps)(MainPanel)
+type Props = {
+  page: Page,
+  selectedTerminalItemId: ?TerminalItemId,
+  ingredSelectionMode: boolean,
+}
 
 function MainPanel(props: Props) {
-  const { page } = props
+  const { page, selectedTerminalItemId, ingredSelectionMode } = props
   switch (page) {
     case 'file-splash':
       return <Splash />
@@ -26,13 +37,34 @@ function MainPanel(props: Props) {
       return <LiquidsPage />
     case 'settings-app':
       return <SettingsPage />
-    default:
-      return <ConnectedDeckSetup />
+    default: {
+      const startTerminalItemSelected =
+        selectedTerminalItemId === START_TERMINAL_ITEM_ID
+      return (
+        <>
+          <MainPageModalPortal>
+            <TimelineAlerts />
+            <Hints />
+            {startTerminalItemSelected && <LabwareSelectionModal />}
+            {!startTerminalItemSelected && <StepEditForm />}
+            {startTerminalItemSelected && ingredSelectionMode && (
+              <LiquidPlacementModal />
+            )}
+          </MainPageModalPortal>
+          <DeckSetup />
+        </>
+      )
+    }
   }
 }
 
 function mapStateToProps(state: BaseState): $Exact<Props> {
   return {
     page: selectors.getCurrentPage(state),
+    selectedTerminalItemId: stepsSelectors.getSelectedTerminalItemId(state),
+    ingredSelectionMode:
+      labwareIngredSelectors.getSelectedLabwareId(state) != null,
   }
 }
+
+export default connect<Props, {||}, _, _, _, _>(mapStateToProps)(MainPanel)
